serverutil/frontend: make cmd/frontend available as a library
This leaves cmd/frontend a simple warpper for the new
serverutil/frontend package.

Also move the templates and images directories from cmd/frontend to doc.
