+++ doc/templates/doc.tmpl
