+++ serverutil/frontend/templates/base.tmpl
@@ -1,0 +1,73 @@
+<!DOCTYPE html>
+<!--
+Copyright 2017 The Upspin Authors. All rights reserved.
+Use of this source code is governed by a BSD-style
+license that can be found in the LICENSE file.
+-->
+<meta charset="utf-8">
+<meta name="viewport" content="width=device-width, initial-scale=1.0">
+<link rel="icon" type="image/png" href="/images/favicon-32x32.png" sizes="32x32">
+<title>{{block "title" .}}Upspin{{end}}</title>
+<style>
+*,
+:before,
+:after {
+  box-sizing: border-box;
+}
+html,
+body {
+  height: 100%;
+}
+body {
+  display: flex;
+  flex-direction: column;
+  font: 14px arial, sans-serif;
+  line-height: 1.45;
+  margin: 0;
+  padding: 0;
+}
+footer {
+  color: #666;
+  font-size: .9em;
+  padding: 1em 1em 2em;
+  text-align: center;
+}
+main {
+  flex: 1;
+  padding: 0 1em;
+}
+.container {
+  margin: 0 auto;
+  max-width: 50em;
+}
+.homeLink:link,
+.homeLink:visited {
+  color: #000;
+  text-decoration: none;
+}
+{{block "styles" .}}{{end}}
+</style>
+<main>
+  <div class="container">
+    {{block "content" .}}Hello, Upspin!{{end}}
+  </div>
+</main>
+<footer>
+  <div>
+    Except as
+    <a href="https://developers.google.com/site-policies#restrictions">noted</a>,
+    the content of this page is licensed under the
+    Creative Commons Attribution 3.0 License,
+    and code is licensed under a
+    <a href="https://github.com/upspin/upspin/blob/master/LICENSE">BSD license</a>.
+  </div>
+  <div>
+    The Upspin mascot is Copyright 2017 Renee French.
+    <a href="https://upspin.io/doc/mascot.md">All Rights
+    Reserved.</a>
+  </div>
+  <div>
+    <a href="/doc/tos.md">Terms of Service</a> |
+    <a href="http://www.google.com/intl/en/policies/privacy/">Privacy Policy</a>
+  </div>
+</footer>
