+++ serverutil/frontend/templates/doc.tmpl
@@ -1,0 +1,20 @@
+<!--
+Copyright 2017 The Upspin Authors. All rights reserved.
+Use of this source code is governed by a BSD-style
+license that can be found in the LICENSE file.
+-->
+{{define "title"}}{{.Title}}{{end}}
+{{define "styles"}}
+img {
+  max-width: 100%;
+}
+table{
+  border-collapse: collapse;
+  width: 100%;
+}
+td {
+  border: 1px solid #666;
+  padding: .25em .45em;
+}
+{{end}}
+{{define "content"}}{{.Content}}{{end}}
